[#10] Minor reorg.
- Moved signature functionality into new `PrivateKey.sign_bundle_at` method.
- Split out `BundleValidator` into own module to avoid circular import.
- Renamed `iota.transaction.{read,write}` to `base` and `creation`, respectively.
- Removed some unit tests that would conflict with multisig (might add back later tho).
