+++ iota/transaction/__init__.py
@@ -2,8 +2,10 @@
 from __future__ import absolute_import, division, print_function, \
   unicode_literals
 
-# Import symbols for backwards-compatibility with PyOTA 1.1.x.
-from .read import *
+# Import symbols to package namespace, for backwards-compatibility with
+# PyOTA 1.1.x.
+from .base import *
+from .creation import *
 from .types import *
 from .utils import *
-from .write import *
+from .validator import *
