+++ iota/transaction/creation.py
@@ -8,11 +8,11 @@ from typing import Iterable, Iterator, List, MutableSequence, Optional, \
 from six import PY2
 
 from iota.crypto import Curl, HASH_LENGTH
-from iota.crypto.signing import KeyGenerator, SignatureFragmentGenerator
+from iota.crypto.signing import KeyGenerator
 from iota.crypto.types import PrivateKey
 from iota.exceptions import with_context
 from iota.json import JsonSerializable
-from iota.transaction.read import Transaction
+from iota.transaction.base import Transaction
 from iota.transaction.types import Fragment, TransactionHash
 from iota.transaction.utils import get_current_timestamp
 from iota.types import Address, Hash, Tag, TryteString
@@ -418,112 +418,7 @@ class ProposedBundle(JsonSerializable, Sequence[ProposedTransaction]):
     if not self.hash:
       raise RuntimeError('Cannot sign inputs until bundle is finalized.')
 
-    # Do lots of validation before we attempt to sign the transaction,
-    # and attach lots of context info to any exception.
-    # This method is likely to be invoked at a very low level in the
-    # application, so if anything goes wrong, we want to make sure it's
-    # as easy to troubleshoot as possible!
-
-    try:
-      txn = self[start_index]
-    except IndexError as e:
-      raise with_context(
-        exc = e,
-
-        context = {
-          'bundle':       self,
-          'key_index':    private_key.key_index,
-          'start_index':  start_index,
-        },
-      )
-
-    # Only inputs can be signed.
-    if txn.value >= 0:
-      raise with_context(
-        exc =
-          ValueError(
-            'Attempting to sign non-input transaction #{i} '
-            '(value={value}).'.format(
-              i     = txn.current_index,
-              value = txn.value,
-            ),
-          ),
-
-        context = {
-          'bundle':       self,
-          'key_index':    private_key.key_index,
-          'start_index':  start_index,
-        },
-      )
-
-    if txn.address.key_index != private_key.key_index:
-      raise with_context(
-        exc =
-          ValueError(
-            'Attempting to sign input transaction #{i} with wrong private key '
-            '(key index: expected={expected}, actual={actual}).'.format(
-              actual    = private_key.key_index,
-              expected  = txn.address.key_index,
-              i         = txn.current_index,
-            ),
-          ),
-
-        context = {
-          'bundle':       self,
-          'key_index':    private_key.key_index,
-          'start_index':  start_index,
-        },
-      )
-
-    if txn.address.security_level != private_key.security_level:
-      raise with_context(
-        exc =
-          ValueError(
-            'Attempting to sign input transaction #{i} with wrong private key '
-            '(security level: expected={expected}, actual={actual}).'.format(
-              actual    = private_key.security_level,
-              expected  = txn.address.security_level,
-              i         = txn.current_index,
-            ),
-          ),
-
-        context = {
-          'bundle':       self,
-          'key_index':    private_key.key_index,
-          'start_index':  start_index,
-        },
-      )
-
-    if txn.signature_message_fragment:
-      raise with_context(
-        exc =
-          ValueError(
-            'Attempting to sign input transaction #{i}, '
-            'but it has a non-empty fragment (is it already signed?).'.format(
-              i = txn.current_index,
-            ),
-          ),
-
-        context = {
-          'bundle':       self,
-          'key_index':    private_key.key_index,
-          'start_index':  start_index,
-        },
-      )
-
-    signature_fragment_generator =\
-      SignatureFragmentGenerator(
-        private_key = private_key,
-        hash_= txn.bundle_hash,
-      )
-
-    # We can only fit one signature fragment into each transaction,
-    # so we have to split the entire signature among the extra
-    # transactions we created for this input in
-    # :py:meth:`add_inputs`.
-    for j in range(txn.address.security_level):
-      self[txn.current_index+j].signature_message_fragment =\
-        next(signature_fragment_generator)
+    private_key.sign_bundle_at(self, start_index)
 
   def _create_input_transactions(self, addy):
     # type: (Address) -> None
